requirements.txt
absl-py==1.4.0
acme==1.21.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
backtrader==1.9.78.123
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.0
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
ckipnlp==1.0.3
ckiptagger==0.2.1
click==8.0.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
contourpy==1.0.7
cryptography==39.0.2
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
duplicity==0.8.21
empyrical==0.5.5
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fasteners==0.14.1
filelock==3.12.2
Flask==2.2.2
flatbuffers==23.5.26
fonttools==4.39.4
fpdf==1.7.2
fredapi==0.5.0
frozendict==2.3.4
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gdown==4.7.1
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
ipykernel==6.23.1
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
josepy==1.10.0
jsonschema==3.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kaleido==0.2.1
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
line-bot-sdk==2.4.1
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
multitasking==0.0.11
mysqlclient==1.4.6
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.24.1
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
paramiko==2.9.3
parsedatetime==2.6
parso==0.8.3
patool==1.12
pdfminer.six==20221105
pdfplumber==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.3
plotly==5.13.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pyfolio==0.9.2
Pygments==2.14.0
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.21.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.1
pytz==2022.7.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
ssh-import-id==5.11
stack-data==0.6.2
systemd-python==234
TA-Lib==0.4.25
tenacity==8.1.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
texttable==1.6.4
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
treelib==1.6.4
trio==0.22.0
trio-websocket==0.9.2
twstock==1.3.1
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
Wand==0.6.11
wcwidth==0.2.6
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.8.2
yfinance==0.2.9
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
